[Reporting/Refactor] Break up screenshots.ts, consolidate logging (#42304)
* [Reporting] Sanitize 409 error log message

* [Reporting] Break up screenshots libs

* Revert "[Reporting] Sanitize 409 error log message"

This reverts commit 61998b5847f5dcbdce98c9db01a9bf1152f254f1.

* [Reporting] Sanitize 409 error log message

* this is for later
