+++ x-pack/legacy/plugins/reporting/export_types/common/constants.ts
@@ -8,3 +8,5 @@ export const LayoutTypes = {
   PRESERVE_LAYOUT: 'preserve_layout',
   PRINT: 'print',
 };
+
+export const WAITFOR_SELECTOR = '.application';
