+++ x-pack/legacy/plugins/reporting/export_types/common/lib/screenshots/types.ts
@@ -1,0 +1,48 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import { ElementPosition } from '../../../../types';
+import { HeadlessChromiumDriver as HeadlessBrowser } from '../../../../server/browsers/chromium/driver';
+import { LayoutInstance } from '../../layouts/layout';
+
+export interface ScreenshotObservableOpts {
+  url: string;
+  conditionalHeaders: any;
+  layout: LayoutInstance;
+  browserTimezone: string;
+}
+
+export interface TimeRange {
+  from: any;
+  to: any;
+}
+
+export interface AttributesMap {
+  [key: string]: any;
+}
+
+export interface ElementsPositionAndAttribute {
+  position: ElementPosition;
+  attributes: AttributesMap;
+}
+
+export interface Screenshot {
+  base64EncodedData: any;
+  title: any;
+  description: any;
+}
+
+export interface ScreenShotOpts {
+  elementsPositionAndAttributes: ElementsPositionAndAttribute[];
+}
+
+export interface BrowserOpts {
+  browser: HeadlessBrowser;
+}
+
+export interface TimeRangeOpts {
+  timeRange: TimeRange;
+}
