+++ x-pack/legacy/plugins/reporting/export_types/png/server/lib/generate_png.js
@@ -16,8 +16,8 @@ function generatePngObservableFn(server) {
 
   const urlScreenshotsObservable = (url, conditionalHeaders, layout, browserTimezone) => {
     return Rx.of(url).pipe(
-      mergeMap(url => screenshotsObservable(url, conditionalHeaders, layout, browserTimezone),
-        (outer, inner) => inner, // eslint-disable-line no-unused-vars
+      mergeMap(url => screenshotsObservable({ url, conditionalHeaders, layout, browserTimezone }),
+        (outer, inner) => inner,
         captureConcurrency
       )
     );
