+++ x-pack/legacy/plugins/reporting/export_types/printable_pdf/server/lib/generate_pdf.js
@@ -33,7 +33,8 @@ function generatePdfObservableFn(server) {
 
   const urlScreenshotsObservable = (urls, conditionalHeaders, layout, browserTimezone) => {
     return Rx.from(urls).pipe(
-      mergeMap(url => screenshotsObservable(url, conditionalHeaders, layout, browserTimezone),
+      mergeMap(
+        url => screenshotsObservable({ url, conditionalHeaders, layout, browserTimezone }),
         captureConcurrency
       )
     );
